+++ frontend/src/App.js
@@ -1,8 +1,11 @@
 import React from "react";
-import "./App.css";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
+
+//******* Import pages *******/
 import Home from "../src/pages/Home";
 import MySearch from "../src/pages/Mysearch";
+import MySearchResult from "../src/pages/MySearchResult";
+import MySearchSubResult from "../src/pages/MySearchSubResult";
 import MyList from "../src/pages/MyList";
 import Community from "../src/pages/Community";
 import SurpriseMe from "../src/pages/SurpriseMe";
@@ -17,6 +20,8 @@ export default function App() {
         <Route path="/" element={<Home />} />
         <Route path="/login" element={<Login />} />
         <Route path="/my-search" element={<MySearch />} />
+        <Route path="/my-search-result" element={<MySearchResult />} />
+        <Route path="/my-search-detail" element={<MySearchSubResult />} />
         <Route path="/my-list" element={<MyList />} />
         <Route path="/community" element={<Community />} />
         <Route path="/surprise-me" element={<SurpriseMe />} />