+++ frontend/webpack.config.js
@@ -1,14 +1,18 @@
 const path = require("path");
 const HtmlWebpackPlugin = require("html-webpack-plugin");
 
- 
-
 module.exports = {
   entry: "./src/index.js",
   output: {
     path: path.resolve(__dirname, "dist"),
     filename: "bundle.js",
     publicPath: "/",
+  },
+  resolve: {
+    extensions: [".js", ".jsx"],
+    alias: {
+      "@src": path.resolve(__dirname, "src"),
+    },
   },
   module: {
     rules: [
@@ -23,16 +27,16 @@ module.exports = {
         },
       },
       {
-        test: /\.(css|sass|scss)$/,
-        use: ["style-loader", "css-loader", "sass-loader"],
-      },
-      {
         test: /\.(png|jpe?g|gif)$/i,
         use: [
           {
             loader: "file-loader",
           },
         ],
+      },
+      {
+        test: /\.css$/,
+        use: ["style-loader", "css-loader"],
       },
     ],
   },
@@ -41,6 +45,8 @@ module.exports = {
       template: path.resolve(__dirname, "public", "index.html"),
     }),
   ],
+
+  // **** Find on stackoverflow ***
   devServer: {
     static: {
       directory: path.join(__dirname, "dist"),
@@ -49,7 +55,7 @@ module.exports = {
     port: 8080,
     historyApiFallback: true,
   },
-};
+}; //******************************* */
 
 //------- Copyright & references ---------------
 //https://legacy.reactjs.org/docs/create-a-new-react-app.html